+++ src/Products/PloneMeeting/tests/testMeeting.py
@@ -1905,12 +1905,36 @@ class testMeetingType(PloneMeetingTestCase):
         self.assertEqual(cfg.getLastMeetingNumber(), 0)
         self.assertEqual(m1.meeting_number, -1)
         self.publishMeeting(m1)
+        # field not used so not initialized
+        self.assertEqual(m1.meeting_number, -1)
+        self.assertEqual(cfg.getLastMeetingNumber(), 0)
+        # now use the field
+        self._enableField('meeting_number', related_to='Meeting')
+        self.backToState(m1, "created")
+        self.publishMeeting(m1)
         self.assertEqual(m1.meeting_number, 1)
         self.assertEqual(cfg.getLastMeetingNumber(), 1)
         m2 = self._createMeetingWithItems()
         self.publishMeeting(m2)
         self.assertEqual(m2.meeting_number, 2)
         self.assertEqual(cfg.getLastMeetingNumber(), 2)
+        # when deleting last meeting, cfg is updated
+        m3 = self._createMeetingWithItems()
+        self.publishMeeting(m3)
+        self.assertEqual(cfg.getLastMeetingNumber(), 3)
+        self.deleteAsManager(m3.UID())
+        self.assertEqual(cfg.getLastMeetingNumber(), 2)
+        # not updated when deleting a meeting in between
+        # but in this case, a warning message is displayed
+        self.deleteAsManager(m1.UID())
+        self.assertEqual(cfg.getLastMeetingNumber(), 2)
+        # when activated, a new year meeting_number is reinit to 1
+        cfg.setYearlyInitMeetingNumbers(('meeting_number', ))
+        next_year = datetime.now().year + 1
+        m4 = self.create('Meeting', date=datetime(next_year, 1, 2))
+        self.publishMeeting(m4)
+        self.assertEqual(m4.meeting_number, 1)
+        self.assertEqual(cfg.getLastMeetingNumber(), 1)
 
     def test_pm_Number_of_items(self):
         '''Tests that number of items returns number of normal and late items.'''
@@ -2854,11 +2878,11 @@ class testMeetingType(PloneMeetingTestCase):
         self.assertEqual(meeting2.get_previous_meeting(), meeting3)
         self.assertFalse(meeting3.get_previous_meeting())
 
-        # very old meeting, previous meeting is searched by default with max 60 days
-        meeting4 = self.create('Meeting', date=meeting3.date - timedelta(days=61))
+        # very old meeting, previous meeting is searched by default with max 180 days
+        meeting4 = self.create('Meeting', date=meeting3.date - timedelta(days=181))
         # still no meeting
         self.assertFalse(meeting3.get_previous_meeting())
-        self.assertEqual(meeting3.get_previous_meeting(interval=61), meeting4)
+        self.assertEqual(meeting3.get_previous_meeting(interval=181), meeting4)
 
     def test_pm_MeetingStrikedAssembly(self):
         """Test use of utils.toHTMLStrikedContent for assembly."""
@@ -2950,24 +2974,36 @@ class testMeetingType(PloneMeetingTestCase):
         # now meeting1 last number is considered 5
         self.assertEqual(unrestricted_view.findFirstItemNumber(), 13)
 
-    def test_pm_FirstItemNumberSetOnClose(self):
-        """First item number is set on closure if it was still -1,
+    def test_pm_FirstItemNumberSetOnDecide(self):
+        """First item number is set when meeting is decided if it was still -1,
            either it is left unchanged."""
         self.changeUser('pmManager')
-        meeting1 = self.create('Meeting')
-        self.assertEqual(meeting1.first_item_number, -1)
-        self.closeMeeting(meeting1)
-        self.assertEqual(meeting1.first_item_number, 1)
-        meeting2 = self.create('Meeting')
-        self.assertEqual(meeting2.first_item_number, -1)
-        self.closeMeeting(meeting2)
-        self.assertEqual(meeting2.first_item_number, 3)
-        # if first_item_number is set, like for example to reinitialize
-        # the first meeting of the year, it is left unchanged
-        meeting3 = self.create('Meeting', first_item_number=1)
-        self.assertEqual(meeting3.first_item_number, 1)
-        self.closeMeeting(meeting3)
-        self.assertEqual(meeting3.first_item_number, 1)
+        m1 = self.create('Meeting')
+        self.assertEqual(m1.first_item_number, -1)
+        self.decideMeeting(m1)
+        # not computed if not used
+        self.assertEqual(m1.first_item_number, -1)
+        # now use the field
+        self._enableField('first_item_number', related_to='Meeting')
+        self.backToState(m1, "created")
+        self.decideMeeting(m1)
+        self.assertEqual(m1.first_item_number, 1)
+        m2 = self.create('Meeting')
+        self.assertEqual(m2.first_item_number, -1)
+        self.decideMeeting(m2)
+        self.assertEqual(m2.first_item_number, 3)
+        # if first_item_number is set it is left unchanged
+        m3 = self.create('Meeting', first_item_number=135)
+        self.assertEqual(m3.first_item_number, 135)
+        self.decideMeeting(m3)
+        self.assertEqual(m3.first_item_number, 135)
+        # first meeting of the year first_item_number may be auto reinit
+        self.meetingConfig.setYearlyInitMeetingNumbers(('first_item_number', ))
+        next_year = datetime.now().year + 1
+        m4 = self.create('Meeting', date=datetime(next_year, 1, 2))
+        self.assertEqual(m4.first_item_number, -1)
+        self.decideMeeting(m4)
+        self.assertEqual(m4.first_item_number, 1)
 
     def test_pm_MeetingAddImagePermission(self):
         """A user able to edit at least one RichText field must be able to add images."""
@@ -3896,6 +3932,7 @@ class testMeetingType(PloneMeetingTestCase):
     def test_pm_Update_first_item_number(self):
         """The the helper Meeting.update_first_item_number that will take in charge"""
         cfg = self.meetingConfig
+        self._enableField('first_item_number', related_to='Meeting')
         cfg.setUseGroupsAsCategories(False)
         self.changeUser('pmManager')
         meeting1 = self._createMeetingWithItems(meetingDate=datetime(2022, 6, 6))